requirements.txt
# Google Cloud dependencies
google-cloud-bigquery==3.17.2
google-auth==2.28.2

# Backend framework
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.10.6
python-multipart==0.0.20

# Database and search
sqlalchemy==2.0.28
psycopg2==2.9.10  # For PostgreSQL
alembic==1.13.1  # For database migrations

# Frontend
jinja2==3.1.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.1
